refactor: optimize integer types and file read limits for memory efficiency
- Change CronSchedule.every_ms type from i64 to u32 for interval storage
- Change HeartbeatService.interval_s type from u64 to u32 for interval storage
- Reduce HEARTBEAT.md file read limit from 1MB to 0.5MB (524288 bytes)
- Reduce config file read limit from 1MB to 0.5MB (524288 bytes)
- Change XevEventLoop delay_ms type from u64 to u32 for timer delay
- Update module documentation comments from /// to //! style in cron.zig

+++ src/agent/cron.zig
@@ -1,6 +1,6 @@
-/// Cron job scheduling module for recurring and one-time tasks.
-/// Jobs can be scheduled to run at specific times or at regular intervals.
-/// Jobs are persisted to disk and survive application restarts.
+//! Cron job scheduling module for recurring and one-time tasks.
+//! Jobs can be scheduled to run at specific times or at regular intervals.
+//! Jobs are persisted to disk and survive application restarts.
 const std = @import("std");
 const Config = @import("../config.zig").Config;
 const Agent = @import("../agent.zig").Agent;
@@ -15,7 +15,7 @@ pub const CronScheduleKind = enum {
 pub const CronSchedule = struct {
     kind: CronScheduleKind,
     at_ms: ?i64 = null, // For 'at' schedules: timestamp in milliseconds
-    every_ms: ?i64 = null, // For 'every' schedules: interval in milliseconds
+    every_ms: ?u32 = null, // For 'every' schedules: interval in milliseconds
 };
 
 /// Payload containing the task to execute and delivery settings.
